Create ListHeader component. There was too much code in List component
